requirements.txt
# The core library for interacting with the Discord API
discord.py>=2.3.0

# Used to load your bot token from the .env file
python-dotenv>=0.21.0

# Recommended library for voice channel functionality, even if not used yet.
# discord.py requires this for voice support.
PyNaCl>=1.5.0